requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
httpx>=0.24.0
beautifulsoup4>=4.12.0
scrapy>=2.8.0
python-dotenv>=1.0.0
firebase-admin>=6.1.0
pydantic>=1.10.7
email-validator>=2.0.0
starlette>=0.26.1
email-validator>=2.0.0
pydantic-settings>=2.0.0